requirements.txt
asgiref==3.5.2
Django==4.1.3
python-dotenv==0.21.0
sqlparse==0.4.3
djangorestframework==3.13.1